refactor: split settler.go
